Added several dumb statements to iterative solvers to resolve issue 146.
Changed Intel compiler configuration for compiling C++ files, now works on a wider range of hardware. Added setting of OMPI_CC, so that mpicc from OpenMPI can choose the proper compiler.

Added PRINT_REAL(x) and PRINT_COMPLEX(x) macros to debug.h (useful for deep numerical debugging). Added cast to int, when printing enum variables (removes warnings on some hardware).

Version incremented to 1.1b7.

+++ trunk/src/chebyshev.c
@@ -57,7 +57,7 @@ static double Zfunc(double x,enum newt_func mode)
 		cn=cos(xn);
 		return (-e*n*sn*c-(1+e*cn)*s)/(2*e*n*sn*s-(1+e*(1+n*n)*cn)*c);
 	}
-	else LogError(ONE_POS,"Unknown mode %d for calling Zfunc",mode);
+	else LogError(ONE_POS,"Unknown mode %d for calling Zfunc",(int)mode);
 }
 
 //==========================================================
@@ -81,7 +81,7 @@ static double Xfunc(double x,enum newt_func mode)
 		cn=cos(xn);
 		return (e*n*cn*c-(1+e*sn)*s)/(-2*e*n*cn*s-(1+e*(1+n*n)*sn)*c);
 	}
-	else LogError(ONE_POS,"Unknown mode %d for calling Zfunc",mode);
+	else LogError(ONE_POS,"Unknown mode %d for calling Zfunc",(int)mode);
 }
 
 //==========================================================
